Collocate single use util
