+++ src/common/utils/fetch-client.ts
@@ -1,5 +1,3 @@
-import { getQueryString } from './query-string';
-
 export class FetchClient {
   constructor(readonly baseURL: string, readonly options?: RequestInit) {
     this.baseURL = baseURL;
@@ -103,4 +101,16 @@ export class FetchError<T> extends Error {
     this.message = message;
     this.response = response;
   }
+}
+
+function getQueryString(queryObj?: Record<string, any>) {
+  if (!queryObj) return undefined;
+
+  const sanitizedQueryObj: Record<string, any> = {};
+
+  Object.entries(queryObj).forEach(([param, value]) => {
+    if (value !== '' && value !== undefined) sanitizedQueryObj[param] = value;
+  });
+
+  return new URLSearchParams(sanitizedQueryObj).toString();
 }
